# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: iHR Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /openai/conversations/{id}/messages:
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a message and stream a response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: {}
...
  /payroll:
    get:
      operationId: listPayslips
      tags: [payroll]
      summary: List payslips
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Payslips list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payslip"
  /payroll/{id}:
    get:
      operationId: getPayslip
      tags: [payroll]
      summary: Get payslip by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payslip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payslip"
...
components:
  schemas:
...
    Payslip:
      type: object
      required: [id, employeeId, month, year, netSalary, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        month:
          type: integer
        year:
          type: integer
        basicSalary:
          type: ["integer", "null"]
        allowances:
          type: ["integer", "null"]
        deductions:
          type: ["integer", "null"]
        netSalary:
          type: integer
        status:
          type: string
          enum: [draft, processed, paid]
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, createdAt]
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, conversationId, role, content, createdAt]
    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required: [title]
    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required: [id, title, createdAt, messages]
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required: [error]